+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/TranslationExtractorPass.php
@@ -11,30 +11,15 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Reference;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\Exception\RuntimeException;
+@trigger_error(sprintf('The %s class is deprecated since version 3.4 and will be removed in 4.0. Use Symfony\Component\Translation\DependencyInjection\TranslationExtractorPass instead.', TranslationExtractorPass::class), E_USER_DEPRECATED);
+
+use Symfony\Component\Translation\DependencyInjection\TranslationExtractorPass as BaseTranslationExtractorPass;
 
 /**
- * Adds tagged translation.extractor services to translation extractor.
+ * Adds tagged translation.formatter services to translation writer.
+ *
+ * @deprecated since version 3.4, to be removed in 4.0. Use {@link BaseTranslationDumperPass instead}.
  */
-class TranslationExtractorPass implements CompilerPassInterface
+class TranslationExtractorPass extends BaseTranslationExtractorPass
 {
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->hasDefinition('translation.extractor')) {
-            return;
-        }
-
-        $definition = $container->getDefinition('translation.extractor');
-
-        foreach ($container->findTaggedServiceIds('translation.extractor', true) as $id => $attributes) {
-            if (!isset($attributes[0]['alias'])) {
-                throw new RuntimeException(sprintf('The alias for the tag "translation.extractor" of service "%s" must be set.', $id));
-            }
-
-            $definition->addMethodCall('addExtractor', array($attributes[0]['alias'], new Reference($id)));
-        }
-    }
 }